fix(main_weathet_model) convert int into double when api return int instead of double

+++ lib/features/current_weather/data/models/main_weather_model.dart
@@ -12,10 +12,10 @@ class MainWeatherModel extends MainWeatherEntity {
 
   factory MainWeatherModel.fromJson(Map<String, dynamic> json) {
     return MainWeatherModel(
-      temp: json["temp"],
-      feelsLike: json["feels_like"],
-      tempMin: json["temp_min"],
-      tempMax: json["temp_max"],
+      temp: json["temp"].toDouble(),
+      feelsLike: json["feels_like"].toDouble(),
+      tempMin: json["temp_min"].toDouble(),
+      tempMax: json["temp_max"].toDouble(),
       pressure: json["pressure"],
       humidity: json["humidity"],
     );
